requirements.txt
flask
itsdangerous
flask-sqlalchemy
passlib
Flask-JWT